+++ WSAdmin/DataModels/TutorsList.swift
@@ -104,9 +104,9 @@ import Foundation
 			let newTutorStudentCount = Int(sheetCells[rowNumber][PgmConstants.tutorStudentCountPosition]) ?? 0
 			let newTutorServiceCount = Int(sheetCells[rowNumber][PgmConstants.tutorServiceCountPosition]) ?? 0
 			let newTutorTotalSessions = Int(sheetCells[rowNumber][PgmConstants.tutorSessionCountPosition]) ?? 0
-			let newTutorCost = Float(sheetCells[rowNumber][PgmConstants.tutorTotalCostPosition]) ?? 0.0
-			let newTutorRevenue = Float(sheetCells[rowNumber][PgmConstants.tutorTotalRevenuePosition]) ?? 0.0
-			let newTutorProfit = Float(sheetCells[rowNumber][PgmConstants.tutorTotalProfitPosition]) ?? 0.0
+			let newTutorCost = Double(sheetCells[rowNumber][PgmConstants.tutorTotalCostPosition]) ?? 0.0
+			let newTutorRevenue = Double(sheetCells[rowNumber][PgmConstants.tutorTotalRevenuePosition]) ?? 0.0
+			let newTutorProfit = Double(sheetCells[rowNumber][PgmConstants.tutorTotalProfitPosition]) ?? 0.0
 			// Create a new Tutor object
 			let newTutor = Tutor(tutorKey: newTutorKey, tutorName: newTutorName, tutorEmail: newTutorEmail, tutorPhone: newTutorPhone, tutorStatus: newTutorStatus, tutorStartDate: newTutorStartDateString, tutorEndDate: newTutorEndDateString, tutorMaxStudents: newTutorMaxStudents, tutorStudentCount: newTutorStudentCount, tutorServiceCount: newTutorServiceCount, tutorTotalSessions: newTutorTotalSessions, tutorTotalCost: newTutorCost, tutorTotalRevenue: newTutorRevenue, tutorTotalProfit: newTutorProfit, timesheetFileID: "")
 			// Add the new Tutor object to the Tutors List object array
