+++ WSAdmin/Views/TutorService.swift
@@ -13,17 +13,17 @@ import Foundation
 	var timesheetServiceName: String
 	var invoiceServiceName: String
 	var billingType: BillingTypeOption
-	var cost1: Float
-	var cost2: Float
-	var cost3: Float
-	var totalCost: Float
-	var price1: Float
-	var price2: Float
-	var price3: Float
-	var totalPrice: Float
+	var cost1: Double
+	var cost2: Double
+	var cost3: Double
+	var totalCost: Double
+	var price1: Double
+	var price2: Double
+	var price3: Double
+	var totalPrice: Double
 	let id = UUID()
 	
-	init(serviceKey: String, timesheetName: String, invoiceName: String,  billingType: BillingTypeOption, cost1: Float, cost2: Float, cost3: Float, price1: Float, price2: Float, price3: Float) {
+	init(serviceKey: String, timesheetName: String, invoiceName: String,  billingType: BillingTypeOption, cost1: Double, cost2: Double, cost3: Double, price1: Double, price2: Double, price3: Double) {
 		self.serviceKey = serviceKey
 		self.timesheetServiceName = timesheetName
 		self.invoiceServiceName = invoiceName
@@ -38,12 +38,12 @@ import Foundation
 		self.totalPrice = price1 + price2 + price3
 	}
 	
-	func computeSessionCostPrice(duration: Int) -> (Float, Float, Float, Float) {
+	func computeSessionCostPrice(duration: Int) -> (Double, Double, Double, Double) {
 		
-		var cost: Float = 0.0
-		var price: Float = 0.0
-		var quantity: Float = 0.0
-		var rate: Float = 0.0
+		var cost: Double = 0.0
+		var price: Double = 0.0
+		var quantity: Double = 0.0
+		var rate: Double = 0.0
 		
 		rate = price1 + price2 + price3
 		if billingType == .Fixed {
@@ -51,7 +51,7 @@ import Foundation
 			cost = cost1 + cost2 + cost3
 			price = rate
 		} else {
-			quantity = Float(duration) / 60.0
+			quantity = Double(duration) / 60.0
 			cost = quantity * cost1 + cost2 + cost3
 			price = quantity * price1 + price2 + price3
 		}
